data-build/requirements.txt
pandas>=2.2.2
numpy==1.26.4
python-dateutil==2.8.2
tqdm>=4.65.0